refactor: confine Tachikoma to a Dashboard submodule; stop the run on quit
Addresses the PR review.

Namespace containment. `using Tachikoma` + `@tachikoma_app` dumped ~150 framework
names into `GNSSReceiver`'s top-level scope, before every `include`. That already
forced `Unitful.linear(...)` where the code used to write bare `linear(...)`, and
the sharper hazard was `view`: `Tachikoma.view` is a brand-new generic, not a
method of `Base.view`, so a plain `view(x, 1:n)` written later anywhere in the
sample path would have failed with a runtime `MethodError`. The dashboard now
lives in `GNSSReceiver.Dashboard` (the layout `@tachikoma_app`'s own docstring
prescribes) and re-exports only `gui`; `src/gui.jl` keeps the data/formatting
layer. `view` at package scope is `Base.view` again and nothing leaks.

Quitting stops the run. `gui` never closed `gui_data_channel`, so `q` returned the
prompt while the feeder, `receive` and the file reader kept churning through the
rest of the recording. `gui` now closes the channel in a `finally`,
`get_gui_data_channel` propagates that close upstream, and `receive` treats a
closed `data_channel` as "consumers are gone" — ending the run quietly instead of
dumping an unhandled-task error as the terminal is handed back — then closes the
sample source it is the sole consumer of.

Panel fixes:

- `_paint_plot!` advances the cursor by `set_string!`'s return value. Recomputing
  it with `textwidth` diverges, since `set_string!` strips escapes and control
  chars, segments graphemes and substitutes a space for a wide glyph straddling
  `max_x`; the `max(1, …)` also shifted a row on a zero-width span. A one-cell
  drift misplaces satellites against the axis labels in the braille sky plot.
- Sky plot puts the zenith at the centre and the horizon on the rim (radius =
  90° − elevation), the standard convention — it was inverted. Rings are labelled
  with elevation as bare numbers, so `°` stays unambiguously an azimuth.
- The PVT panel no longer claims it is "re-acquiring" after the stream has ended,
  which contradicted the header.
- `Esc` no longer quits: undocumented, and far too easy to hit by reflex.
- CN0 panel title carries its unit again (`[dBHz]`), lost with the old `xlabel`.
- Dropped the `pvt !== nothing` checks — `GUIData.pvt` is not a `Union{…,Nothing}`
  — and the `get_LLA` try/catch in the Location panel, which only blanked that
  panel while the PVT panel above called the same conversion unguarded.
- Documented the `9`-column chrome allowance in the CN0 bar width.

Tests: the stale-header path is now asserted positively (a fix with
`pvt_fresh = false`), the held-`last_fix` branch of the PVT panel is covered in
both its wordings, and the new teardown is verified end to end.
