+++ test/gui.jl
@@ -30,6 +30,34 @@
     @test isnothing(gui_datas[1].pvt.time)
 end
 
+@testset "Closing the GUI channel stops the pipeline" begin
+    sat_data_type = GNSSReceiver.SatelliteDataOfInterest{SVector{2,ComplexF64}}
+    # An endless producer, standing in for a receiver replaying a long recording.
+    data_channel = Channel{GNSSReceiver.ReceiverDataOfInterest{sat_data_type}}() do ch
+        runtime = 0.0u"s"
+        while true
+            runtime += 1.0u"s"
+            put!(
+                ch,
+                GNSSReceiver.ReceiverDataOfInterest{sat_data_type}(
+                    Dictionary{Tuple{Symbol,Int},sat_data_type}(),
+                    GNSSReceiver.PVTSolution(),
+                    runtime,
+                ),
+            )
+        end
+    end
+
+    gui_data_channel = get_gui_data_channel(data_channel, 200ms)
+    take!(gui_data_channel)     # one frame reaches the GUI
+    close(gui_data_channel)     # ... and then the user quits with `q`
+
+    # `gui` closes its channel on exit, and the forwarding task propagates that close
+    # upstream, so the producer is shut down instead of grinding on with nobody watching.
+    timedwait(() -> !isopen(data_channel), 10.0)
+    @test !isopen(data_channel)
+end
+
 # Render one frame of the Tachikoma dashboard for `model` into an off-screen buffer and
 # return it as plain text. This drives the real `view` (CN0 bars, sky plot, PVT block,
 # location block) without a TTY.
@@ -39,36 +67,74 @@ function render_gui_text(model; width = 140, height = 50)
     area = Rect(1, 1, width, height)
     buf = Buffer(area)
     frame = Frame(buf, area, GraphicsRegion[], Tuple{Int,Int,Matrix{ColorRGBA}}[])
-    GNSSReceiver.view(model, frame)
+    GNSSReceiver.Dashboard.view(model, frame)
     buffer_to_text(buf, area)
 end
 
-function gui_model(gui_data; show_diagnostics = false)
-    m = GNSSReceiver.ReceiverModel()
+function gui_model(gui_data; show_diagnostics = false, last_fix = gui_data)
+    m = GNSSReceiver.Dashboard.ReceiverModel()
     m.gui = gui_data
-    isnothing(gui_data.pvt.time) || (m.last_fix = gui_data)
+    isnothing(last_fix.pvt.time) || (m.last_fix = last_fix)
     m.show_diagnostics = show_diagnostics
     m
 end
 
+# A minimal `GUIData` carrying a real fix — four GPS L1 satellites plus a position/velocity/
+# time solution — for the header- and held-fix tests below.
+function fixed_gui_data(; runtime = 10.0u"s", pvt_fresh = true)
+    sat_data_type = GNSSReceiver.SatelliteDataOfInterest{SVector{2,ComplexF64}}
+    sat_keys = [(:GPSL1CA, 3), (:GPSL1CA, 12), (:GPSL1CA, 23), (:GPSL1CA, 10)]
+    GNSSReceiver.GUIData(
+        Dictionary(
+            sat_keys,
+            [sat_data_type(45.0dBHz, zeros(SVector{2,ComplexF64}), true) for _ in sat_keys],
+        ),
+        PositionVelocityTime.PVTSolution(;
+            position = ECEF(4.0e6, 3.9e5, 4.9e6),
+            velocity = ECEF(2.0e6, 2.9e5, 1.9e6),
+            time = TAIEpoch(2022, 10, 8),
+            dop = PositionVelocityTime.DOP(1.0, 1.0, 1.0, 1.0, 1.0),
+            sats = Dictionary(
+                sat_keys,
+                [
+                    PositionVelocityTime.SatInfo(
+                        ECEF(5e6 + 1e5 * i, 3e6 - 1e5 * i, 1e6 + 2e5 * i),
+                        0.0,
+                        0.0u"m",
+                    ) for i in eachindex(sat_keys)
+                ],
+            ),
+        ),
+        runtime,
+        pvt_fresh,
+    )
+end
+
 @testset "GUI input handling (update!)" begin
     using Tachikoma: KeyEvent
+    using GNSSReceiver.Dashboard: ReceiverModel, update!
 
-    # Quit on q / Q / Esc.
-    for e in (KeyEvent('q'), KeyEvent('Q'), KeyEvent(:escape))
-        m = GNSSReceiver.ReceiverModel()
-        GNSSReceiver.update!(m, e)
+    # Quit on q / Q / Ctrl-C — the keys the docs advertise.
+    for e in (KeyEvent('q'), KeyEvent('Q'), KeyEvent(:ctrl_c))
+        m = ReceiverModel()
+        update!(m, e)
         @test m.quit
     end
 
+    # Esc is deliberately *not* bound to quit: it is far too easy to hit by reflex to tear
+    # down a running receiver with.
+    m = ReceiverModel()
+    update!(m, KeyEvent(:escape))
+    @test !m.quit
+
     # Diagnostics are shown by default; `d` toggles them; an unbound key is a no-op.
-    m = GNSSReceiver.ReceiverModel()
+    m = ReceiverModel()
     @test m.show_diagnostics
-    GNSSReceiver.update!(m, KeyEvent('d'))
+    update!(m, KeyEvent('d'))
     @test !m.show_diagnostics
-    GNSSReceiver.update!(m, KeyEvent('d'))
+    update!(m, KeyEvent('d'))
     @test m.show_diagnostics
-    GNSSReceiver.update!(m, KeyEvent('x'))
+    update!(m, KeyEvent('x'))
     @test !m.quit && m.show_diagnostics
 end
 
@@ -85,6 +151,37 @@ end
     out = render_gui_text(m)
     @test occursin("stream ended", out)
     @test !occursin("stale", out)
+end
+
+@testset "GUI header flags a stale fix" begin
+    # A fix is on screen, but the receiver produced no new one for this frame
+    # (`pvt_fresh = false`) and the stream is still running: that — and only that — is
+    # "stale". The stream-ended case above must not use this wording.
+    out = render_gui_text(gui_model(fixed_gui_data(; pvt_fresh = false)))
+    @test occursin("stale (no new fix)", out)
+    @test !occursin("stream ended", out)
+end
+
+@testset "GUI holds the last fix when the current frame has none" begin
+    fix = fixed_gui_data(; runtime = 10.0u"s")
+    # Current frame carries no fix (`pvt.time === nothing`) but an earlier one did, so the
+    # PVT panel falls back to the held fix and flags why it is frozen.
+    searching =
+        GNSSReceiver.GUIData(fix.sat_data, GNSSReceiver.PVTSolution(), 20.0u"s", false)
+    m = gui_model(searching; last_fix = fix)
+
+    out = render_gui_text(m)
+    @test occursin("last fix (re-acquiring)", out)
+    # The held fix's coordinates stay on screen, stamped with the *current* run time.
+    @test occursin("Coordinates: 50.830895°N, 5.568737°E", out)
+    @test occursin("Run time:", out) && occursin("20.0 s", out)
+
+    # Once the stream has ended nothing is being re-acquired any more — the panel must not
+    # claim otherwise while the header says "stream ended".
+    m.stream_ended = true
+    out = render_gui_text(m)
+    @test occursin("last fix (stream ended)", out)
+    @test !occursin("re-acquiring", out)
 end
 
 @testset "GUI with no data" begin
